GitBotAI/code_generation.py
```python
import os
from GitBotAI.templates import code_templates

class CodeGenerator:
    def __init__(self, github_token):
        self.github_token = github_token

    def identify_missing_files(self, repo_name):
        # This function will interact with GitHub's API to identify missing files
        # This is a placeholder and needs to be implemented with actual logic
        pass

    def generate_code(self, missing_files):
        generated_code = {}

        for file in missing_files:
            if file in code_templates:
                generated_code[file] = code_templates[file]
            else:
                # Use machine learning model to generate code
                # This is a placeholder and needs to be implemented with actual logic
                generated_code[file] = "Generated code"

        return generated_code

    def ensure_compatibility(self, repo_name, generated_code):
        # This function will ensure that the generated code is compatible with the existing codebase
        # This is a placeholder and needs to be implemented with actual logic
        pass

    def update_repo(self, repo_name, generated_code):
        # This function will update the GitHub repository with the generated code
        # This is a placeholder and needs to be implemented with actual logic
        pass

    def generate_missing_code(self, repo_name):
        missing_files = self.identify_missing_files(repo_name)
        generated_code = self.generate_code(missing_files)
        self.ensure_compatibility(repo_name, generated_code)
        self.update_repo(repo_name, generated_code)
```
