GitBotAI/tests/test_code_generation.py
```python
import unittest
from GitBotAI import code_generation

class TestCodeGeneration(unittest.TestCase):

    def setUp(self):
        self.code_gen = code_generation.CodeGeneration()

    def test_identify_missing_code(self):
        repo = "test_repo"
        missing_files = self.code_gen.identify_missing_code(repo)
        self.assertIsInstance(missing_files, list)

    def test_generate_code(self):
        missing_file = "test_file.py"
        generated_code = self.code_gen.generate_code(missing_file)
        self.assertIsInstance(generated_code, str)

    def test_check_compatibility(self):
        repo = "test_repo"
        missing_file = "test_file.py"
        generated_code = self.code_gen.generate_code(missing_file)
        is_compatible = self.code_gen.check_compatibility(repo, generated_code)
        self.assertIsInstance(is_compatible, bool)

if __name__ == '__main__':
    unittest.main()
```